requirements.txt
# DeepLense GSoC 2026 Evaluation Pipeline Dependencies

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Astrophysics & Physics Engines
lenstronomy>=1.11.0
astropy>=5.3.0

# Data Manipulation & Computation
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning Metrics & Preprocessing
scikit-learn>=1.3.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing & Utilities
Pillow>=9.5.0
tqdm>=4.65.0